(#181) Rework variable_length_btree-test.c to work w/ CTests
This commit refactors and ports the existing stand-alone
variable_length_btree-test.c to run under the CTests framework.

- Break-out common definitions and some inline functions that
  are shared between .c and test.c code into new private
  header file, variable_length_btree_private.h
- Refactor static inline fns from .c to _private.h to get debug builds working.

Two new files are created for those test cases:
  - variable_length_btree_test.c
  - variable_length_btree_stress_test.c

No new test-code logic is added / changed. Existing test assertions
are converted to use CTests-style assertion checking.

Shared setup code is refactored into btree_test_common.c/.h
Delete older form of standalone unit/variable_length_btree-test.c
and rules in Makefile to build that binary.

- Some static inline fns from .c that are a "bit too big" are
  moved to extern functions. Add prototypes in
  variable_length_btree_private.h
