+++ Makefile
@@ -19,7 +19,6 @@ INCDIR               = include
 
 SRC := $(shell find $(SRCDIR) -name "*.c")
 FUNCTIONAL_TESTSRC := $(shell find $(FUNCTIONAL_TESTSDIR) -name "*.c")
-UNITSRC := $(shell find $(UNITDIR) -name "*.c")
 UNIT_TESTSRC := $(shell find $(UNIT_TESTSDIR) -name "*.c")
 
 OBJ := $(SRC:%.c=$(OBJDIR)/%.o)
@@ -28,16 +27,17 @@ OBJ := $(SRC:%.c=$(OBJDIR)/%.o)
 FUNCTIONAL_TESTOBJ= $(FUNCTIONAL_TESTSRC:%.c=$(OBJDIR)/%.o)
 
 # Objects from unit-test sources in tests/unit/ sub-dir
-# Resolves to a list: obj/tests/unit/a.o obj/tests/unit/b.o ...
+# Resolves to a list: obj/tests/unit/a.o obj/tests/unit/b.o obj/tests/unit/c.o
 UNIT_TESTOBJS= $(UNIT_TESTSRC:%.c=$(OBJDIR)/%.o)
 
+# ----
 # Binaries from unit-test sources in tests/unit/ sub-dir
 # Although the sources are in, say, tests/unit/kvstore_basic_test.c, and so on ...
 # the binaries are named bin/unit/kvstore_basic_test (Drop the 'tests'.)
-# Resolves to a list: bin/unit/a bin/unit/b ...
-UNIT_TESTBINS= $(UNIT_TESTSRC:tests/%.c=$(BINDIR)/%)
-
-UNITBINS= $(UNITSRC:%.c=%)
+# Also, there may be other shared .c files that don't yield a standalone
+# binary. Hence, only build a list from files named *_test.c
+# Resolves to a list: bin/unit/a_test bin/unit/b_test bin/unit/c_test ...
+UNIT_TESTBINS= $(UNIT_TESTSRC:tests/%_test.c=$(BINDIR)/%_test)
 
 # Automatically create directories, based on
 # http://ismail.badawi.io/blog/2017/03/28/automatic-directory-creation-in-make/
@@ -107,11 +107,11 @@ endif
 #*********************************************************#
 # Targets to track whether we have a release or debug build
 #
-all: $(LIBDIR)/libsplinterdb.so $(LIBDIR)/libsplinterdb.a $(BINDIR)/driver_test $(UNITBINS) \
+all: $(LIBDIR)/libsplinterdb.so $(LIBDIR)/libsplinterdb.a $(BINDIR)/driver_test \
         unit_test $(EXTRA_TARGETS)
 
 # Any libraries required to link test code will be built, if needed.
-tests: $(BINDIR)/driver_test $(UNITBINS) unit_test
+tests: $(BINDIR)/driver_test unit_test
 
 release: .release all
 	rm -f .debug
@@ -150,10 +150,16 @@ $(BINDIR)/driver_test : $(FUNCTIONAL_TESTOBJ) $(LIBDIR)/libsplinterdb.so | $$(@D
 	$(LD) $(LDFLAGS) -o $@ $^ $(LIBS)
 
 # Target will build everything needed to generate bin/unit_test along with all
-# the individual binaries for each unit-test case.
+# the individual binaries for each unit-test case. Most unit-tests are fairly
+# well-contained, but variable_length_btree_test needs several objects to be
+# linked with it, which will be produced while building the shared library.
 $(BINDIR)/unit_test: unit_test
 unit_test: $(UNIT_TESTBINS) $(LIBDIR)/libsplinterdb.so
-	$(LD) $(LDFLAGS) -o $(BINDIR)/$@ $(UNIT_TESTOBJS) $(LIBDIR)/libsplinterdb.so $(LIBS)
+	$(LD) $(LDFLAGS) -o $(BINDIR)/$@ $(UNIT_TESTOBJS)                   \
+                            $(OBJDIR)/tests/functional/test_data.o      \
+                            $(OBJDIR)/src/platform_linux/platform.o     \
+                            $(LIBDIR)/libsplinterdb.so $(LIBS)
+
 
 $(LIBDIR)/libsplinterdb.so : $(OBJ) | $$(@D)/.
 	$(LD) $(LDFLAGS) -shared -o $@ $^ $(LIBS)
@@ -185,21 +191,6 @@ $(OBJDIR)/%.o: %.c | $$(@D)/.
 # List the individual unit-tests that can be run standalone and are also
 # rolled-up into a single unit_test binary.
 
-$(OBJDIR)/unit/variable_length_btree-test.o: src/variable_length_btree.c
-unit/variable_length_btree-test: $(OBJDIR)/tests/functional/test_data.o     \
-                                 $(OBJDIR)/src/util.o                       \
-                                 $(OBJDIR)/src/data_internal.o              \
-                                 $(OBJDIR)/src/mini_allocator.o             \
-                                 $(OBJDIR)/src/rc_allocator.o               \
-                                 $(OBJDIR)/src/config.o                     \
-                                 $(OBJDIR)/src/clockcache.o                 \
-                                 $(OBJDIR)/src/platform_linux/platform.o    \
-                                 $(OBJDIR)/src/task.o                       \
-                                 $(OBJDIR)/src/platform_linux/laio.o        \
-                                 $(OBJDIR)/src/platform_linux/platform.o
-	mkdir -p $(BINDIR)/unit;
-	$(LD) $(LDFLAGS) -shared $^ -o $(BINDIR)/$@
-
 # ---- main() is needed to drive each standalone unit-test binary.
 $(BINDIR)/unit/main: $(OBJDIR)/tests/unit/main.o
 $(OBJDIR)/tests/unit/main.o: tests/unit/main.c
@@ -214,9 +205,39 @@ unit/kvstore_basic_test: $(OBJDIR)/tests/unit/kvstore_basic_test.o        \
 
 # ----
 $(BINDIR)/unit/kvstore_basic_stress_test: unit/kvstore_basic_stress_test
-unit/kvstore_basic_stress_test: $(OBJDIR)/tests/unit/kvstore_basic_stress_test.o       \
-                                $(OBJDIR)/tests/unit/main.o                      \
+unit/kvstore_basic_stress_test: $(OBJDIR)/tests/unit/kvstore_basic_stress_test.o    \
+                                $(OBJDIR)/tests/unit/main.o                         \
                                 $(LIBDIR)/libsplinterdb.so
+	mkdir -p $(BINDIR)/unit;
+	$(LD) $(LDFLAGS) -o $(BINDIR)/$@ $^ $(LIBS)
+
+# ----
+# String together all objects needed to link variable_length_btree test binaries
+VARIABLE_LENGTH_BTREE_TEST_OBJS = $(OBJDIR)/tests/unit/btree_test_common.o      \
+                                  $(OBJDIR)/tests/functional/test_data.o        \
+                                  $(OBJDIR)/src/util.o                          \
+                                  $(OBJDIR)/src/data_internal.o                 \
+                                  $(OBJDIR)/src/mini_allocator.o                \
+                                  $(OBJDIR)/src/rc_allocator.o                  \
+                                  $(OBJDIR)/src/config.o                        \
+                                  $(OBJDIR)/src/clockcache.o                    \
+                                  $(OBJDIR)/src/variable_length_btree.o         \
+                                  $(OBJDIR)/src/platform_linux/platform.o       \
+                                  $(OBJDIR)/src/task.o                          \
+                                  $(OBJDIR)/src/platform_linux/laio.o
+
+$(BINDIR)/unit/variable_length_btree_test: unit/variable_length_btree_test
+unit/variable_length_btree_test: $(OBJDIR)/tests/unit/variable_length_btree_test.o  \
+                                 $(OBJDIR)/tests/unit/main.o                        \
+                                 $(VARIABLE_LENGTH_BTREE_TEST_OBJS)
+	mkdir -p $(BINDIR)/unit;
+	$(LD) $(LDFLAGS) -o $(BINDIR)/$@ $^ $(LIBS)
+
+# ----
+$(BINDIR)/unit/variable_length_btree_stress_test: unit/variable_length_btree_stress_test
+unit/variable_length_btree_stress_test: $(OBJDIR)/tests/unit/variable_length_btree_stress_test.o  \
+                                 $(OBJDIR)/tests/unit/main.o                        \
+                                 $(VARIABLE_LENGTH_BTREE_TEST_OBJS)
 	mkdir -p $(BINDIR)/unit;
 	$(LD) $(LDFLAGS) -o $(BINDIR)/$@ $^ $(LIBS)
